requirements.txt
pandas>=1.3.0
scikit-learn>=1.0.0
nltk>=3.6.0
wordfreq>=2.5.0
pyyaml>=6.0
